refactor: update mining difficulty type from int to int32 for consistency
**Changed:**

- Changed `Difficulty` field in `ServerMining` struct from `int` to `int32` for type consistency with blockchain logic in `config/config.go`
- Updated blockchain initialization in `server/main.go` to pass `cfg.Mining.Difficulty` directly as `int32` without conversion

+++ server/main.go
@@ -163,7 +163,7 @@ func checkTLSCertificates(cfg *config.ServerConfig) {
 
 func initializeBlockchainAndPool(cfg *config.ServerConfig) (*Blockchain, *MiningPool) {
 	fmt.Printf("Initializing blockchain with difficulty: %d\n", cfg.Mining.Difficulty)
-	blockchain := NewBlockchain(int32(cfg.Mining.Difficulty))
+	blockchain := NewBlockchain(cfg.Mining.Difficulty)
 	fmt.Println("Initializing mining pool...")
 	pool := NewMiningPool(blockchain)
 	return blockchain, pool
